Update crypto configs and test

+++ lib/src/test/java/datagram/SharedConfigTest.java
@@ -150,7 +150,7 @@ public class SharedConfigTest {
             sequenceNumberField.setAccessible(true); // Make it accessible
 
             // Set sequence number to simulate a message replay attack
-            sequenceNumberField.setInt(secureSocketBase, 0);
+            sequenceNumberField.setLong(secureSocketBase, 0L);
 
             clientSocket.send(sendPacket);
         }
